Created equipment system

+++ Assets/Scripts/Items/Firearm.cs
@@ -5,9 +5,9 @@ namespace Assets.Scripts
     public abstract class Firearm : Weapon
     {
         [SerializeField] [Range(1, 100)] protected int _capacity;
-        [SerializeField] [Range(0, 5)] int _reloaringTime;
+        [SerializeField] [Range(0, 5)] float _reloaringTime;
 
         public int Capacity => _capacity;
-        public int ReloadingTime => _reloaringTime;
+        public float ReloadingTime => _reloaringTime;
     }
 }
